Add base10 flag to nsimplify
The default behavior is to convert floats to rationals using the string
representation (roughly). Setting this flag to false will cause nsimplify to
use the exact binary representation (i.e., it round-trips with evalf(), given
the correct precision).

I personally believe the default for this flag should be False, but I've left
that alone for now. See #12045.