tests: drop erf from pickling test
Well... because it makes integrals from its "as_integral" property
and those integrals have a Dummy variable of integration.  When they
are unpickled, the "as_integral" properties are different (because its
a new Dummy variable).  Related to #5539 whereby integrals with same
integrands and limits but different variables of integration are not
equal.