error fcns: test rewrite('Integral') works with deep=True
Its important that `_eval_rewrite_as_Integral(self, *args)` uses `*args*`
and not self when doing the rewriting.  The difference is important for
composition (where `self.args` will be `z` whereas `*args` may contain
a written child from the composition).  One or more of my previous drafts
had this wrong, so add a test which protects this case.