physics.hydrogen: Added normalized hydrogen wave functions psi_{nlm}
New feature:
In addition to the already existing radial wave functions of the
hydrogen atom, now there's also the possibility to get the full
wavefunction psi_nlm, depending on the quantum numbers n, l and m
and the spherical coordinates r, phi and theta. Check Wikipedia
"Hydrogen atom" or any quantum text book for validation of the
formula.

See docstring for examples.