Add __eq__ to PolyRing (based on __hash__)
The added test does not pass because the dtype of a PolyElement a dynamically
created class, which doesn't have any metaclass (hence, it uses the default
__eq__). But that is set to PolynomialRing.dtype, and is used in the
PolynomialRing.__eq__ test. I am not sure how to properly fix this. It seems
odd that the dtype should be created dynamically at each class creation, and
also odd that PolynomialRing.__eq__ is doing an equality check on a class.