Reduce hyperfunctions with no shifts
The algorithm applied by hyperexpand for finding representations
of generalized hypergeometric functions is based on shift
operators [Roach1996] in order to reduce the number of parameters.
Shift operators are differential operators that apply to each
term of the series individually.

If any of the lower parameters of a hypergeometric series is
a nonpositive integer, then only a finite number of the
coefficients have nonzero denominator. Then the series can be
meaningfully defined only if the numerators are also zero so that
the series becomes a polynomial.

In the polynomial case, the shift operators apply only to terms
with nonzero denominators. If the reduction by shift
operators leads to a well defined series with no zeros in any of
the denominators, the result may not be valid in general.
This commit will avoid shift operators is those situations.
The only reduction is elimination of common upper and lower
parameters.