a Number should know its own floor/ceiling
Knowledge about floors and ceilings of Numbers was storied in the
Functions floor and ceiling.  Instead, we make each subclass of Number
have a .floor() and .ceiling() method.

In the older code, the standalone floor/ceiling Functions were written
to assume all non-Rational non-Float Numbers were Integers (which
might not be true if other Number subclasses were introduced in the
future).  Also, its better OO practice for Rationals to know their own
floor and ceiling.