matrices: LU decomposition of empty matrix doesn't raise error
When the input matrix has a zero length dimension,
LUdecomposition_simple() now returns a matrix of zeros of the same
size, and an empty list of row permutations.
The previous behavior was to raise a ValueError.

Example where number of columns in A is zero, and both
A.LUdecomposition_Simple() and A.LUdecomposition()
return an answer:
>>> A
Matrix(3, 0, [])
>>> lu, p = A.LUdecomposition_Simple()
>>> lu
Matrix(3, 0, [])
>>> p
[]
>>> L, U, p = A.LUdecomposition()
>>> L
Matrix([
[1, 0, 0],
[0, 1, 0],
[0, 0, 1]])
>>> U
Matrix(3, 0, [])
>>> p
[]
>>> L*U
Matrix(3, 0, [])

Cosmetic changes:
Reduced the number of blank lines down to a level that could
arguably be described as sparingly.
Replaced lu.shape[1] with lu.cols.