matrices: Added pivot search strategy to LU decomposition functions
LUdecomposition_Simple() and LUdecomposition() now accept a keyword
argument that specifies a callable that returns the location of the
next pivot. The default behavior is to raise a ValueError if the
rank of the matrix being decomposed is strictly less than the number
of rows and columns. This mimics the behavior of the previous
implementation of LUdecomposition_Simple().
The new helper function _next_pivot_down() searches
only in the column below the current pivot location, and raises a
ValueError if no suitable pivot can be found.
The helper function _next_pivot_down_right() searches the entire
submatrix below and to the right of the current pivot position,
and signals that no pivot was found by returning the dimensions
of the matrix being decomposed.

Both LUdecomposition_Simple() and LUdecomposition() respect the
iszerofunc keyword argument.