ComplexFloat: comment about print-level formatting of Float
Behaviour of zeros in Float printing depends on whether the object is
by itself ("1.2500000000", "2.") or in an expression ("1.25", "2.0").
Note sure yet which we want for ComplexFloat but this commit adds some
tests that I worked out while understanding the difference.  Adds a
comment to the printing code about how the Float-like behaviour can
be achieved.

Also update the docstrings for expr.py for ComplexFloat printing;
these are subtle (its where I noticed this issue) and might need
updating if the printer behaviour changes.