Dummy: generate dummy_index sequentially from a random base value
Instead of randomly generating dummy_index and keeping them in dict
to keep track of order, we generate dummy_index sequentially from a
random starting value.  This guarantees that Dummys created in the
usual way ("d = Dummy()") are unique and distinct from one another.
If any Dummy has a dummy_index specified, then there is a extremely
small chance that a subsequent "d = Dummy()" could match.  The random
starting value lowers the chance of this occurring.