v2pt_theory: Added zero velcoity constratint requirement.
I'm unsure if what I am adding here, which is a trivial detail, was
intentionally left out but here is the description:
P and O are two points on a rigid body B that is moving in frame N.
The velcoity of P in N, N^v^P, can be computed using v2pt_thepry,
which it does admirably. However, the function should also set
B^v^P = 0; that is what I have added. In fact, the function should
likely also constrain the B^v^O to be 0; that is what makes P and O
2 points of the same body.

Further, this constraint of B^v^O should also likely be added to v1pt
theory. If you agree, I can commit changes to that next time, as well.