add number_hasimaginary and number_isnonzero to geometry/util.py
The ideal fix is to get the query system to evaluate numers when
making a decision about properties like zero. For ask(Q.zero(add))
the Add handler is triggered and if that doesn't decide then the
Expr handler is triggered...but that one uses the old assumptions.
The numerical handling in the Basic handler is not run.