enabled matching of is_Pow expressions with is_Mul
Previously, the following program returned None:
```
In [1]: b = Wild('b',exclude=[x])

In [2]: m = Wild('m',exclude=[x])

In [3]: expr = (4*x)**2

In [4]: expr.match((b*x)**m)
```
because the class of expression is Mul and class of pattern is Pow. To
solve this problem, I used `expand_power_base()` to convert the pattern
into a Mul expression and matched it with the expression.

Tests are added.