solveset(log(x)/x <= 0, x, S.Reals) ->(0, oo) instead of (0, 1]: #12429
I tried to fix this bug. I think that the problem is in the line 104 of solveset.py.
Sympy can't invert the function log(x)/x. In this case the method returns as 's' the value '0',
i.e. the rhs of equation log(x)/x=0. So s.intersection(domain) has not sense in this situation.
I added a test to check whether the result is an inversion. Probably it is not the best way,
but it seems to work.