Convergent series: improved a test for using the integral method
In is_convergent() method, sometimes the condition
'is_decreasing(sequence_term, interval)'
for using the integral test returns (correctly) 'False' and the integral test for
convergence is ignored.
The condition for using the integral test is that 'sequence_term' is definitely decreasing
(i.e. for a certain point).
So I tried to implement this condition.
Now 'Sum(log(n)/n**3, (n, 1, oo)).is_convergent()' returns 'true'. This is due
to the fact that 'log(n)/n**3' is not decreasing in (1, oo), but it definitely
decreasing.