Test for convergence of series added
A test for a sequence of the type
Sum(a_n * b_n)
where a_n (or b_n) is bounded and Sum(b_n) (or Sum(a_n)) is convergent.
In such a case the original is convergent, by a comparison principle.
Added also some tests of the type
Sum(sin(n)/n**3)