solve: removed redundant Dummy vars and subexpressions as symbols
The Dummy variables that used to be added to enable `solve(3)` to
behave like `solve(3,x)` are unnecessary and caused complications
in other situations. Additionally, when extracting symbols from
expressions, there is no need to add subexpressions involving the
free symbols to the symbols list (especially since there was no check
to see if the values of subexpressions in the solution list are
consistent with the values of the free symbols) unless it is
an application of an undefined function which could be useful to
solve for.