matrices: LU partial reversion to naive pivot search strategy
The doctests timeout when the LU decomposition functions use
matrices._find_reasonable_pivot() as the pivot search.
Future versions of the LU decomposition functions can pass
the keyword arguments iszerofunc and simpfunc to a future
versionof _find_reasonable_pivot() that completes in less
time.
The current pivot search algorithm employed by
LUdecomposition_simple() first looks for a pivot that iszerofunc
guarantees is nonzero. If no such candidate exists,
the candidates are first simplified if the caller does not pass
simpfunc=None, and then a second pivot search more permissive pivot
search is employed. In the second search, the pivot is the first
candidate that iszerofunc can guarantee is nonzero.
If no such candidate pivot exists, then the pivot is the first
candidate for which iszerofunc returns None.
If no such candidate exists, the pivot search starts again in the
column to the right.