Add safe flag to sympify()
For now, True (the default) causes it to raise UnsafeSympifyError on strings
with attribute access.

We may also want to add other things. For instance, when we add support for
f-strings, they will need to be added. It may also be judicious to remove
syntactic elements that we don't use, like comprehensions, though I'm unsure
if they can really result in unsafe code without attribute access.

Fixes #10805.