print Interval with methods indicating openess
It is perhaps more salient to see Intervals printed
with an explicit word telling the closure rather than
having to remember that True or False refer to the
interval being open or closed, respectively:

Interval.Ropen(3, 4) == Interval(3, 4, False, True)

The form on the left is now the default str form that is shown.

When Infinity is involved, it is not explicitly indicated
as open since neither oo nor -oo are included in intervals, e.g.

Interval.Ropen(3, oo) ==> Interval(3, oo)