Fix issue_12555. limit((3**x + 2 * x**10) / (x**10 + E**x), x, -oo) gives 0 instead of 2
The wrong behavior is due to the fact that leading terms of
numerator and denominator was wrong calculated.
At some point the substitution x = 1/u was performed and then
the leading terms of numerator and denominator
were calculated comparing the order of the addends. In the
case of the numerator we have O(3**(1/u)) and O(u**(-10)).
When comparing the previous order, sympy assumes that u tends to 0, but
positive. Instead in this situation u should be negative.