integrals: Add new trigonometric rule in manual integration
A new trigonometric rule is added which can recognise sin(2*x)
and replace it with 2*sin(x)*cos(x) based on the double angle formula.
This change makes the result of integrate(sin(2*x)*cos(x)) to have a much
simpler form and may provide an alternate strategy for solving complex
trigonometric integrals similar to problem in issue #12641.

Examples:

>>> from sympy import integrate, sin, cos, pi
>>> from sympy.abc import x
>>> integrate(sin(2*x))
-cos(2*x)/2
>>> integrate(cos(x)*sin(2*x), manual=True)
-2*cos(x)**3/3
>>> integrate((sin(2*x)*cos(x))/(1+cos(x)), manual=True)
-2*log(cos(x) + 1) - cos(x)**2 + 2*cos(x)
>>> integrate((sin(2*x)*cos(x))/(1+cos(x)), (x, 0, pi/2), manual=True)
-1 + 2*log(2)