Fix issue 12677
This fix concerns issue 12677, where the definite integral
integrate(sin(x)/(cos(x)**3), (x, 0, pi/6))
In _eval_interval(self, x, a, b) of expr.py, sympy calculates the singularities
of the primitive, as the zeros of the denominator. This returns in an infinite set; hence
there is an infinite loop. Intersecting the singularities with the domain should avoid
this situation.

Added a test for the issue.