Change internal representations from dict to tuple/list
The dictionary representations for best origin and direction
parameter(w.r.t hyperplane) are changed to tuple/list form.

This is because of discrepancies arising due to {x: a, y: b}
and {y: b, x: a} referring to the same object. This change may
be temporary or permanent depending on how scalable and robust it
proves to be for the final 2D API and the 3D use case.