modify _eval_power methods
o The holonomic test was modified because the string was failing
at the insignificant digits of the imaginary part (which is
just supposed to be small).

o The apart test was modified to pass

in master apart(f) gave I/((2 + I)*(x + I)) - 2*(-2 + I)/(5*(x - 2))
and now this becomes (in each step of the test):

>>> apart(f)
-I*(-2 + I)/(5*(x + I)) - 2*(-2 + I)/(5*(x - 2))
>>> together(_)
(-2/5 + I/5)*(-2*x - I*(x - 2) - 2*I)/((x - 2)*(x + I))
>>> factor(_)
-x*(-2 + I)*(2 + I)/(5*(x - 2)*(x + I))

o The rewriting of sqrt(1/i) in the test,

    i = symbols('i', imaginary=True)
    # this expansion of sqrt(1/i) happens at line 297 of power.py
    assert sqrt(3/i) == Mul(sqrt(3), sqrt(-i)/abs(i), evaluate=False)

happens in line 296 of power.py. If this is not desired, the condition
"and b.is_number" could be added to the end of line 295.

            # Please enter the commit message for your changes. Lines starting