octave: don't print newlines between rows in matrices
Maybe this could be re-introduced only for "big" matrices and only
when generating m-files.  But octave_code itself is often used to
generate single-line expressions where the newline is at best a
distraction.  Even in generated m-files the previous behaviour did
not look good because the subsequent lines were not indented properly.