Revert to standard algorithm and use dynamic for only multiple case.
On further testing the standard algorithm performed faster than the
dynamic one. However, if the terms of the polynomial are such that
every monomial's partial derivatives are present previously then dynamic
is faster. Therefore, it is used for the multiple case, that is when a
maximum degree is supplied and the values for a list of polynomials is to
be calculated.