Allow for complex Numpy scalars in sympify
As of Sympy 1.1, the following raises a recursion error:

>>> c = np.longcomplex(1 + 1j)
>>> print(sympify(c))

This is because the sympify code had the following:

   if not isinstance(a, np.floating):
       return sympify(np.asscalar(a))

In the case of `c` above, it is not `np.floating` type (it's complex)
and `np.asscalar(c)` returns another numpy scalar (rather than a Python
scalar, as would be the case for the `np.complex` type), so causing the
loop to go into infinite recursion.

Add handling of the complex case directly in the sympify clause above,
and add test for `np.longcomplex` type.

While I'm at it, add test for `np.longdouble` type, which will be
`np.float96` on 32-bit platforms, `np.float128` on 64-bit.