simplify Relational.canonical
The previous canonical function did not make a canonical
expression and something like x**2 > -y/x would oscillate
in form with successive applications of the method.

>>> r = x**2 > -y/x
>>> r.canonical == r.canonical.canonical
False

Now a much simpler, easily verifiable function is presented.

The goal is not simplification -- that can be done in
an _eval_simplify routine. Once little simplifications
are started, it is easy to make the function do too much.
e.g. if the sign is going to be cleared, then why not the
leading coefficient, why not use _invert and clear as
much as possible, etc....