Fix various recursion issues in Add assumptions
These only show up when running with the cache off, as otherwise the caching
prevents infinite recursion.

Unfortunately, I don't know how to test this, aside from running the test
suite with the cache off (which we should be doing on Travis or somewhere but
aren't). There are also a lot of other assumedly unrelated test failures
with ./bin/test -C at the moment, although you can see an improvement as
'./bin/test -C core' crashes without this patch and doesn't with it.

Fixes #12907.