modify piecewise_fold
The tests are added in such a way that the 3 regions that are produced
in 10087 are tested to be correct. It's a matter of style whether 3 or
more pieces are returned. Currently, a redundant condition is returned
since no simplification occurs during the folding. Such simplification
would be worked out when needed, e.g. during integration which requires
the identification of intervals upon which the function is defined.