Fix implemented_function for latest changes to UndefinedFunction
This is a somewhat backwards incompatible change. It no longer modifies
UndefinedFunction objects inplace. With UndefinedFunction now actually
defining __eq__ and __hash__, it should no longer be considered mutable.
Instead, if implemented_function is passed an UndefinedFunction it creates a
new UndefinedFunction with the same name with the implemented attached.