integrals: fixed bug which produced a wrong result when integrating derivatives
Fixes issue #12899. The original logic was faulty in several ways:
a) treating a derivative with respect to a variable different
from integration symbol as if it was constant.
b) looking only at the last variable of the differentiation.
Fixed a test in test_manual. Added new tests in test_manual
This makes a mild assumption that integration and differentiation commute.