Fixed arbitrary_point() bug
Originally, the arbitrary_point() function used undirected Segments to calculate the points. This assumption sometimes did not work, as the order of points used to create the Polygon might be inverted, thus returning the wrong value. This fix uses Rays instead of Segment objects. Rays are directed lines, thus, the order of Points used to create the Polygon is maintained. Another change necessary so the rest of the code base remains intact: create a new @sides_rays property, which is analogous to the @sides property but returns Rays instead of Segments.