Simplification with assumptions: refine and abs
This deals with issue 12724. In case of 'refine(abs(x*y), Q.positive(x))'
the result was 'Abs(x*y)', which is mathematically correct. A better result
is 'x * Abs(y)'.

This PR proposes a change in 'refine_abs()' method in 'sympy/assumptions/refine.py'.
In the case the arg is a multiplication, the Abs is splitted according to the factors of arg.
A test is added in 'sympy/assumptions/tests/test_refine.py'.