Remove type sorting in meijerint _mytype
Sortability of SymPy types is going to be removed. The stated purpose of the
object is to create a hash representing the type of f. For general objects,
the args cannot be sorted anyway. For commutative objects, like Add and Mul,
the args are already sorted, and the sort key is based on the type (this could
change, but for now it is). So the sorting seems unnecessary.