fix iteration over rules in make_confluent
The list of rules is extended in the double loop in make_confluent
but the for loop ignored the new elements. This is fixed now.
Additionally, _reduce_redundacies was run too early in the loop
and would sometimes lead to a dictionary key error. This is also
taken care of now.