Use frozenset in meijerint _mytype
The previous behavior of sorting needed to change (sorting of types is going
away). However, not sorting is also wrong, as we need the keys to be
canonicalized. Since the order does not appear to matter (the "type" is just
used as a quick lookup before doing a full match), just use a frozenset.

This commit was originally from #13059.