rewriting system changes and comments
add_rule method changed to directly add all rules that can be obtained from
a given pair of words (it wasn't a major issue as the other rules
would eventually be found by make_confluent but it is probably more
efficient this way). Overlaps in make_confluent should be found with
the inverses of the rules as well as their regular versions. The change
in the reduce method was made because the previous "if" condition could
lead to problems (e.g. with a hypothetical rule b -> a). symbols_index
method for free groups was only used in one place in the code and didn't
do what intented (because negative of 0 is 0) so it is removed and the
__le__ method of FreeGroupElements is fixed. The other changes
are to do with shortening the code or adding comments.