Don't use as_ordered_terms/factors in sort_key
These functions call evalf to order terms numerically, which is too slow for a
sort key. The sort key is used in the core Add and Mul ordering, among other
places, where we don't care about numeric ordering. We just want something
consistent. But it does need to be as fast as possible.