_solve_inequality mods
The condition that a be non-zero requires checking whether it
is not False; if it is True is *is* zero and if it is None it
might be zero -- in both cases we do not divide by a.

When an expression is solved in the "linear" section of the code
we still should record the conditions under which the solution
is valid. Right now we just require that denominators be
non-zero. Any denominator that disappeared with simplification
needs to have its condition registered as part of the return
value, e.g. x < x*(1/x + 1) -> x < 1 + x which is True...
but if x == 0 then the relationship is not valid and we return
False.