could_extract_minus_sign is sign invariant aware
docstring and function have been modified to indicate
that for any foo where foo = -foo, could_extract_minus_sign
will return False. 0 and a Mul with a factor of zoo are two
examples of things that behave this way.

Negation of a Mul could be modified so that the first Add factor
could take the negation, but this is a separate issue. The added
test will continue to pass even if Mul is changed that way.