better generators for alt and change alt check
This commit rearranges and modifies the code related to generators
of alternating groups so that they are consistent with those
desribed in the docstring. Also, the check for whether a group is
alternating is replaced by considering its order: checking if (0 1)
is contained in the group takes far too long for large degrees.