periodicity: added support for modulo operation
periodicity(f, symbol) uses decompogen to decompose a function, and
uses recursion until one of the function in decompogen in (f,symbol)
has its periodicity implemented as a base case. As of this commit, Mod
function is one of the base cases, so periodicity of modulo function can be
evaluated as well.

Example:
In [1]: periodicity(tan(3*x%4),x)
Out[1]: 4/3

In [2]: periodicity(tan((x**2)%3),x)

In [3]: periodicity(sin(x)%3,x)
Out[3]: 2⋅π