polys: Norm of polynomial over a number field
SymPy implements `sqf_norm`, the norm of a modified polynomial,
for the construction of primitive elements of number fields.
It is the same as `norm` if that is square-free but different
in general. This PR implements the standard norm in all cases.
The low-level function, `dmp_norm`, is in sqfreetools.py because
it is essentially a single step of `dmp_sqf_norm`.