Removed some redundant checks from isprime
Numbers removed from a list used with an `in` test. The list is composite numbers that are exceptions to a power and modulo test. The removed numbers are divisible by previously tested values hence will not reach this point (confirmed `[n for n in range(2, 23002) if isprime(n)]` remained unchanged).