agca: Finite ring extensions
Let A be a (commutative) ring and B an extension ring of A.
An element t of B is a generator of B (over A) if all elements
of B can be represented as polynomials in t with coefficients
in A. The representation is unique if and only if t satisfies no
non-trivial polynomial relation, in which case B can be identified
with a (univariate) polynomial ring over A.

The polynomials having t as a root form a non-zero ideal in general.
The most important case in practice is that of an ideal generated by
a single monic polynomial. If t satisfies such a polynomial relation,
then its highest power t^n can be written as linear combination of
lower powers. It follows, inductively, that all higher powers of t
also have such a representation. Hence the lower powers t^i
(i = 0, ..., n-1) form a basis of B, which is then called a finite
extension of A, or, more precisely, a monogenic finite extension
as it is generated by a single element t.