move _symbol/_uniquely_named_symbol to symbols.py
charpoly and gauss_jordan_solve will no longer produce
expressions with clashing symbols

The generator of the charpoly is recast to a nonclashing symbol
since the PurePoly instance does not depend on what symbol is
used for the generator. This only means that care needs to be taken
when wanting to use the generator as in `is_nilpotent`. (There,
a uniquely named symbol was *passed* as the generator; if that wasn't
done then the `x` used in the return value would have to be
p.gen as in `if p.args[0] == p.gen ** self.rows`.)