manualintegrate modifications
Since some tests are done on the values obtained from the
match, it is important that S.Zero, not 0, be taken as the
default if the value is not present.

Also, in the loop over expr, the Wilds a and b were being
redefined as the values. Now A and B are the symbols and
a and b are the retrieved values. Although this completes
one more cycle through the loop (correctly) it doesn't lead
to a different result in the test that traverses this condition
but a different case might succeed where the former code would
fail.