Restrict functions for mathematica parsing
(From #13533)

Problem:
So far, there is no input check phase, hence all the functions in
the input are just let go with lowercase.
For example, spell-missed 'Arcsin[1/2]' is parsed to 'pi/6'.

Improved:
I set up a barricade, and only registered functions in dictionary
are able to be parsed. It is a large restriction, but it would be
safer. There are two good reasons for supporting this.

(1) We can exclude invalid functions mentioned above.

(2) Not all functions with the same name between Mathematica and
    sympy do also the same work.

    Example) Log[2, 4] = 2 (Mathematica)
	     log(4, 2) = 2 (sympy)

    Base positions are different!

    So, we really need to check a function one by one about how it
    works in Mathematica and sympy. Then parse it not only the name,
    but also the movement.

Above the reasons, we should restrict available function in parsing.

Future task:
(1) Increase the number of available functions.

(2) In the present, list and matrix is not taken into account. My code
    throws a ValueError if it finds '{' or '}'. This point should be
    improved.