Wrong result with solveset. Issue 13350
Fix 13350.
The expression solveset(x**2 - 2*x - 15, symbol = x, domain = Interval(-oo, 0))
returns EmptySet() instead of FiniteSet(-3). The bug is due to the
function _invert(f_x, y, x, domain=S.Complexes). It returns
x1, s.intersection(domain)
but in such example x1 is x**2 - 2*x, so that 's' does not represent
the solution. Hence the intersection with the domain is meaningless.

In the case x1 is not equal to the symbol, it is better to return x1, s

A test is added.