_verify and strong presentation
Implementation of Sims Verify algorithm and a function for computing
strong presentations of permutation groups. This is more efficient
than obtaining a strong presentation by just using `presentation` on
the strong generators. `_verify` can be eventually used to obtain a
base and a strong generating set for a group by generating a candidate
randomly and amending missing generators which would be faster than
Schreier-Sims for large groups.