last argument of _rs_series() changed to prec
https://github.com/sympy/sympy/blob/master/sympy/polys/ring_series.py#L1913

On changing the last argument of the function _rs_series() in the above line to just prec the issue disappears for both python 2 and 3.
The argument (prec - sum_pows +min_pows[i]) was just to reduce further calculations(thus reducing the load on rs_trunc().
But in this particular issue, the expansion of exp(a + b**QQ(6,7)) comes out to be exp(b**(6,7)), while it should be a*exp(b**(6,7)) + exp(b**(6,7)), therefore reducing a term from the series.