geometry: make arbitrary_point of Plane run through all points
Changes Plane.arbitrary_point so that it is able to represent
any point in the plane, by way of linear equations with two parameters
u and v (these appear to be traditional names of surface parameters, as
opposed to t for curve parameter). Closes #13799.

The method random_point is changed accordingly. Incidentally, it was
found that a plane could previously be created with zero normal vector.
A check and a test are added to prevent this.