directed segments, self-intersecting poly
Segments now retain the order of their args like all
other linear entities. This required some attention
for computing projections since the order of the
points computed in sets is not canonical. There is an
awareness of this issue for Partition -- see line 1515
in sets.py. The same could be done for GeometryEntity
as an alternative to the changes here.

Also added a smallest_angle_between method to compute
the smallest angle between two lines as opposed to the
non-reflex (oriented) angle between lines as is computed
in the angle_between method.

smallest_angle_between -> 0 <= angle <= pi/2
angle_between -> 0 <= angle < pi

The reflect method for RegularPentagon was corrected
and another test added (and the former incorrect test
updated). The former test gave points that were very
close to correct but were not exactly right when checked
by equals as is done now.