Correct the docstring of as_independent method of expressions
The docstring claims that the second part of the tuple returned by
as_independent(self, *deps, **hint)
"will be 1 or else have terms that contain variables that are in deps".
This is only true if self is a Mul. For an Add in the same
situation the second term will be 0 rather than 1.

For example, (x+y).as_independent(z) returns (x + y, 0)