last stabilizer isn't always cyclic + block action fix
The last stabilizer in the chain isn't always cyclic and this
commit accounts for it. Also, the case in the strong_presentation
algorithm where the relators are found in the block action was
slightly corrected as the domain of one of the homomorphism should
be the whole group, otherwise composition doesn't work properly.

Additionally a new test was added and `uniq` replaced by `set`
everywhere in the new functions.