Improve the solver for linear 1st order ODE systems of size 2
This commit improves the handling of linear ODE systems
x' = a*x + b*y
y' = c*x + d*y
with constant coefficients. The formulas are cleaned up and rewritten
in terms of eigenvectors. Particular attention is paid to complex and
symbolic coefficients, and to some edge cases not previously covered.