Correctly dispatch the printing of custom vs built-in functions
Undefined functions (whether applied or not) are not longer sent
to custom _print_foo methods designed for printing built-in functions.
Additionally, correct printing for beta function is added for both
LaTeX and pretty print (this function incorrectly printed as
lowercase beta, in all printers). The logic for LaTeX printing of
un-applied built-in functions like gamma or DiracDelta is moved
from assorted classes to LaTeX printer. The same logic is now applied
to pretty printer as well. Fixes #2832 and miscellaneous minor bugs.