Enable Piecewise to be plotted: fixing experimental_lambdify fallback
Experimental_lambdify attempts to use python_cmath first, casting the arguments
as complex. If this leads to an error due to inequality comparison with complex
type, it falls back on real arguments and python_math. However there was a bug
where only errors due to '>' comparison led to fallback, while '<' was fatal.

For example,
```
plot(Piecewise((x, x < 1), (x**2, True)))
```
now works (it fails in current master, at least on Python 3.6).

A test is added. The only previous test for Piecewise plotting had > sign in it
but no < sign, which is why it passed.