Do not apply expand_complex to expressions with AppliedUndef
`solveset` sometimes sends solutions to expand_complex in order to
separate the real and imaginary parts of numeric answers. Since this
only complicates the expressions with symbols, there is a check
```
expr.free_symbols == set()
```
However this check does not detect the presence of applied undefined
functions such as f(0), which also make the complex expansion
complicated. So the check was changed to
```
expr.atoms(Symbol, AppliedUndef) == set()
```
Fixes #13825