Only use inferred assumptions if the integral did not evaluate
In some cases, imposing a correct assumption (such as "x is real" when
integrating from -oo to oo) makes things works: instead of quickly
failing on a wrong path, the integrator persists and takes very long
or/and returns a needlessly complicated result. So let's not do this
unless we have to.