Correct casting of NumPy floats to SymPy floats
Use the correct value of binary precision: np.finfo(x).nmant + 1
The tests in test_number did not detect the error because they
were not implemented correctly either.

Sympify interprets np.finfo(x).nmant as binary precision, which is
wrong. nmant is the number of bits used to store the mantissa,
for example 52 in double precision. But this does not count the
leading bit of mantissa, which is always 1 and therefore not stored.
The binary precision, in the sense of the parameter used by SymPy and
mpmath, is 53 for double precision floats, not 52.

As a result of this discrepancy, sympify loses 1 bit of precision for
NumPy types float16 and float32 (not for default float, because for
that one it uses hardcoded 53). Examples:
```
np.float16(2/3)            #  0.6665
srepr(S(np.float16(2/3)))  #  Float('0.66602', precision=10)")
```
and

```
int(np.float32(2/3) * 2**24)     # 11184811
int(S(np.float32(2/3)) * 2**24)  # 11184812
```

The tests in test_number did not detect this because they had
np.float16(2)/3 and np.float32(2)/3
In the process of division by 3, NumPy upcasts these floats, so both
of these fractions are identical to each other and to np.float64(2)/3
The correct way is test is with np.float16(2/3) and np.float32(2/3)